requirements.txt
asgiref==3.4.1
dj-static==0.0.6
Django==3.2.6
django-cors-headers==3.12.0
djangorestframework==3.13.1
gunicorn==20.1.0
pytz==2021.1
sqlparse==0.4.1
static3==0.7.0
tzdata==2022.1
